# tests/webapp/openapi_phase_a.expected.json
{
  "components": {
    "schemas": {
...
      "Body_post_step2_onboarding_step2_post": {
        "properties": {
          "starting_capital": {
            "title": "Starting Capital",
            "type": "string"
          },
          "universe": {
            "title": "Universe",
            "type": "string"
          }
        },
        "required": [
          "starting_capital",
          "universe"
        ],
        "title": "Body_post_step2_onboarding_step2_post",
        "type": "object"
      },
      "Body_post_step3_onboarding_step3_post": {
        "properties": {
          "strategy": {
            "title": "Strategy",
            "type": "string"
          }
        },
        "required": [
          "strategy"
        ],
        "title": "Body_post_step3_onboarding_step3_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionRequest": {
        "description": "Body for ``POST /api/session`` \u2014 the operator pastes a token\nissued via the operator-tooling pipeline. ``account_id`` is\n``\"household\"`` for read-scope sessions or the targeted account\nid for mutation-scope sessions.",
        "properties": {
          "account_id": {
            "default": "household",
            "description": "claim the token SHALL carry (defaults to the household sentinel)",
            "title": "Account Id",
            "type": "string"
          },
          "token": {
            "description": "operator HMAC token",
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "FastAPI surface for the trading-bot. Phase A ships the live-state read + registry-promotion mutation + HTMX dashboard. Phase B adds SSE + JobQueue + cookie sessions.",
    "title": "trading-bot webapp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/session": {
      "delete": {
        "operationId": "close_session_api_session_delete",
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        },
        "summary": "Close the cookie session"
      },
      "post": {
        "description": "Verifies the supplied operator token and sets the trading-session cookie (HttpOnly + Secure + SameSite=Strict). Subsequent GETs may omit the Authorization header.",
        "operationId": "open_session_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open a cookie-backed session"
      }
    },
...
    "/onboarding/step2": {
      "post": {
        "description": "POST /onboarding/step2 \u2014 validate step 1, advance to step 2.",
        "operationId": "post_step2_onboarding_step2_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_step2_onboarding_step2_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Step2"
      }
    },
    "/onboarding/step3": {
      "post": {
        "description": "POST /onboarding/step3 \u2014 validate step 2, advance to step 3.",
        "operationId": "post_step3_onboarding_step3_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_step3_onboarding_step3_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Step3"
      }
    },
    "/paper-sessions/{account_id}/stop": {
      "post": {
        "description": "Stop the paper-trading runtime keyed on ``account_id``.\n\nAuth: any valid token claim (household OR per-account) is\naccepted \u2014 the dashboard view uses the same gate. Mutation\nis idempotent: stopping a non-existent / already-stopped\nsession is a no-op (the redirect still lands the operator\nback on the dashboard).",
        "operationId": "post_stop_paper_sessions__account_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Stop"
      }
    },
...
  }
}